Attempting to move from Lark 0.6.7 to 0.7.0...
(Cherry-picked from a commit of 2 Mar 2022, on an earlier version of
the pfsc-server repo.)

Not all unit tests are passing yet, but this seems like a reasonable start.
*Some* unit tests have been repaired by the changes made so far.

It seems there were major changes when Lark moved from 0.6.7 to 0.7.0.
At least the lexer changed, but other things seem to have changed as well.

We're no longer allowed to use weights on terminals. If we drop the weights,
unit tests break. If we want to keep them, we can say `lexer='standard'`, but then
the unit tests break in a different way.

Someone opened an issue <https://github.com/lark-parser/lark/issues/383> on this
topic, and was advised that you can still use weights on *non*terminals. So we're trying
that here. It seems to have solved some of the problems, but not others.

The `common.lark` grammar also changed in 0.7.0. Specifically, the change in the definition of
`ESCAPED_STRING` seems to have effectively shut off the `DOTALL` regex flag, i.e. it seems that
escaped strings are no longer allowed to contain newlines. So instead of importing `ESCAPED_STRING`
from `common.lark`, we just replicate the old definition from 0.6.7 (and call it `QUOTE_STRING`).

Even with all these changes, still many unit tests are crashing, and I can't give this any more
time right now.